+++ src/extensions/sentry-integration.ts
@@ -18,6 +18,7 @@
 
 import { Properties } from '../types'
 import { PostHog } from '../posthog-core'
+import { ErrorProperties } from './exceptions/error-conversion'
 
 // NOTE - we can't import from @sentry/types because it changes frequently and causes clashes
 // We only use a small subset of the types, so we can just define the integration overall and use any for the rest
@@ -40,6 +41,15 @@ interface _SentryIntegration {
     setupOnce(addGlobalEventProcessor: (callback: _SentryEventProcessor) => void, getCurrentHub: () => _SentryHub): void
 }
 
+interface SentryExceptionProperties {
+    $sentry_event_id: any
+    $sentry_exception: any
+    $sentry_exception_message: any
+    $sentry_exception_type: any
+    $sentry_tags: any
+    $sentry_url?: string
+}
+
 export class SentryIntegration implements _SentryIntegration {
     name: string
 
@@ -55,19 +65,28 @@ export class SentryIntegration implements _SentryIntegration {
             addGlobalEventProcessor((event: _SentryEvent) => {
                 if (event.level !== 'error' || !_posthog.__loaded) return event
                 if (!event.tags) event.tags = {}
+
                 const host = _posthog.config.ui_host || _posthog.config.api_host
                 event.tags['PostHog Person URL'] = host + '/person/' + _posthog.get_distinct_id()
                 if (_posthog.sessionRecordingStarted()) {
                     event.tags['PostHog Recording URL'] = _posthog.get_session_replay_url({ withTimestamp: true })
                 }
+
                 const exceptions = event.exception?.values || []
-                const data: Properties = {
+
+                const data: SentryExceptionProperties & ErrorProperties = {
+                    // PostHog Exception Properties,
+                    $exception_message: exceptions[0]?.value,
+                    $exception_type: exceptions[0]?.type,
+                    $exception_personURL: host + '/person/' + _posthog.get_distinct_id(),
+                    // Sentry Exception Properties
                     $sentry_event_id: event.event_id,
                     $sentry_exception: event.exception,
                     $sentry_exception_message: exceptions[0]?.value,
                     $sentry_exception_type: exceptions[0]?.type,
                     $sentry_tags: event.tags,
                 }
+
                 if (organization && projectId)
                     data['$sentry_url'] =
                         (prefix || 'https://sentry.io/organizations/') +