+++ src/posthog-core.ts
@@ -55,6 +55,8 @@ import {
 import { SentryIntegration } from './extensions/sentry-integration'
 import { createSegmentIntegration } from './extensions/segment-integration'
 import { PageViewIdManager } from './page-view-id'
+import { ExceptionObserver } from './extensions/exceptions/exception-autocapture'
+import { ErrorEventArgs } from './extensions/exceptions/error-conversion'
 
 /*
 SIMPLE STYLE GUIDE:
@@ -103,6 +105,8 @@ const defaultConfig = (): PostHogConfig => ({
     ui_host: null,
     token: '',
     autocapture: true,
+    // TODO: change to undefined when we release this so that remote config can turn it on even if unconfigured by the user
+    autocapture_exceptions: false,
     rageclick: true,
     cross_subdomain_cookie: document?.location?.hostname?.indexOf('herokuapp.com') === -1,
     persistence: 'cookie',
@@ -222,6 +226,9 @@ const create_phlib = function (
     instance.webPerformance = new WebPerformanceObserver(instance)
     instance.webPerformance.startObservingIfEnabled()
 
+    instance.exceptionAutocapture = new ExceptionObserver(instance)
+    instance.exceptionAutocapture.startObservingIfEnabled()
+
     instance.__autocapture = instance.get_config('autocapture')
     autocapture._setIsAutocaptureEnabled(instance)
     if (autocapture._isAutocaptureEnabled) {
@@ -275,6 +282,7 @@ export class PostHog {
     toolbar: Toolbar
     sessionRecording: SessionRecording | undefined
     webPerformance: WebPerformanceObserver | undefined
+    exceptionAutocapture: ExceptionObserver | undefined
 
     _captureMetrics: CaptureMetrics
     _requestQueue: RequestQueue
@@ -769,6 +777,23 @@ export class PostHog {
      */
     push(item: SnippetArrayItem): void {
         this._execute_array([item])
+    }
+
+    /*
+     * PostHog supports exception autocapture, however, this function
+     * is used to manually capture an exception
+     * and can be used to add more context to that exception
+     *
+     * Properties passed as the second option will be merged with the properties
+     * of the exception event.
+     * Where there is a key in both generated exception and passed properties,
+     * the generated exception property takes precedence.
+     */
+    captureException(exception: Error, properties?: Properties): void {
+        this.exceptionAutocapture?.captureException(
+            [exception.name, undefined, undefined, undefined, exception],
+            properties
+        )
     }
 
     /**