+++ src/types.ts
@@ -53,6 +53,8 @@ export interface PostHogConfig {
     ui_host: string | null
     token: string
     autocapture: boolean | AutocaptureConfig
+    // this is currently in development and may have breaking changes without a major version bump
+    autocapture_exceptions: boolean
     rageclick: boolean
     cross_subdomain_cookie: boolean
     persistence: 'localStorage' | 'cookie' | 'memory' | 'localStorage+cookie' | 'sessionStorage'
@@ -143,6 +145,7 @@ export interface SessionRecordingOptions {
     collectFonts?: boolean
     inlineStylesheet?: boolean
     recorderVersion?: 'v1' | 'v2'
+    recordCrossOriginIframes?: boolean
 }
 
 export enum Compression {
@@ -202,6 +205,8 @@ export interface DecideResponse {
     errorsWhileComputingFlags: boolean
     autocapture_opt_out?: boolean
     capturePerformance?: boolean
+    // this is currently in development and may have breaking changes without a major version bump
+    autocaptureExceptions?: boolean
     sessionRecording?: {
         endpoint?: string
         consoleLogRecordingEnabled?: boolean